# front-end/generate/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Android Device Farm",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/accounts": {
      "get": {
        "tags": ["accounts"],
        "summary": "List Accounts Endpoint",
        "description": "List accounts owned by the current user with optional filters.",
        "operationId": "list_accounts_endpoint_api_accounts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountOut"
                  },
                  "title": "Response List Accounts Endpoint Api Accounts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["accounts"],
        "summary": "Create Account Endpoint",
        "description": "Create a new account. Password is encrypted before storage.",
        "operationId": "create_account_endpoint_api_accounts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/import": {
      "post": {
        "tags": ["accounts"],
        "summary": "Bulk Import Json",
        "description": "Bulk import accounts from a JSON array. Duplicate (platform, username) pairs are skipped.",
        "operationId": "bulk_import_json_api_accounts_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkImportBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/accounts/import-csv": {
      "post": {
        "tags": ["accounts"],
        "summary": "Bulk Import Csv",
        "description": "Stream-import accounts from a CSV file upload.\n\nReads the upload in 64 KB chunks, parses CSV incrementally, and flushes\nbatches of up to 500 rows to the DB via INSERT ON CONFLICT DO NOTHING —\nso memory usage stays flat regardless of file size.\n\nExpected CSV columns: platform, username, password, display_name, tags, notes",
        "operationId": "bulk_import_csv_api_accounts_import_csv_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_bulk_import_csv_api_accounts_import_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/accounts/round-robin": {
      "post": {
        "tags": ["accounts"],
        "summary": "Round Robin Assign Endpoint",
        "description": "Auto-assign accounts to devices in round-robin order.",
        "operationId": "round_robin_assign_endpoint_api_accounts_round_robin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoundRobinBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Round Robin Assign Endpoint Api Accounts Round Robin Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/accounts/{account_id}": {
      "get": {
        "tags": ["accounts"],
        "summary": "Get Account Endpoint",
        "description": "Get account details including device links.",
        "operationId": "get_account_endpoint_api_accounts__account_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountWithLinksOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["accounts"],
        "summary": "Update Account Endpoint",
        "description": "Update account fields. Password (if provided) is re-encrypted before storage.",
        "operationId": "update_account_endpoint_api_accounts__account_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["accounts"],
        "summary": "Delete Account Endpoint",
        "description": "Delete account and all its device links (cascade).",
        "operationId": "delete_account_endpoint_api_accounts__account_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{account_id}/status": {
      "patch": {
        "tags": ["accounts"],
        "summary": "Update Account Status",
        "description": "Manually set account status (active / banned / cooldown / disabled).",
        "operationId": "update_account_status_api_accounts__account_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{account_id}/devices": {
      "get": {
        "tags": ["accounts"],
        "summary": "List Account Devices Endpoint",
        "description": "List all device-account links for an account.",
        "operationId": "list_account_devices_endpoint_api_accounts__account_id__devices_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceAccountOut"
                  },
                  "title": "Response List Account Devices Endpoint Api Accounts  Account Id  Devices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["accounts"],
        "summary": "Assign Device To Account",
        "description": "Assign a device to an account.",
        "operationId": "assign_device_to_account_api_accounts__account_id__devices_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDeviceBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceAccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{account_id}/devices/{device_id}": {
      "delete": {
        "tags": ["accounts"],
        "summary": "Unassign Device From Account",
        "description": "Remove a device-account link.",
        "operationId": "unassign_device_from_account_api_accounts__account_id__devices__device_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountCreate": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "default": ""
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "tags": {
            "type": "string",
            "title": "Tags",
            "default": ""
          },
          "proxy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Id"
          },
          "account_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Account Metadata",
            "default": {}
          }
        },
        "type": "object",
        "required": ["platform", "username"],
        "title": "AccountCreate"
      },
      "AccountOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "cooldown_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldown Until"
          },
          "proxy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Id"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "tags": {
            "type": "string",
            "title": "Tags"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "total_usage_minutes": {
            "type": "number",
            "title": "Total Usage Minutes"
          },
          "usage_today_minutes": {
            "type": "number",
            "title": "Usage Today Minutes"
          },
          "usage_reset_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage Reset Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "platform",
          "username",
          "display_name",
          "status",
          "cooldown_until",
          "proxy_id",
          "notes",
          "tags",
          "user_id",
          "created_at",
          "updated_at",
          "last_used_at",
          "total_usage_minutes",
          "usage_today_minutes",
          "usage_reset_date"
        ],
        "title": "AccountOut"
      },
      "AccountStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": ["status"],
        "title": "AccountStatusUpdate"
      },
      "AccountUpdate": {
        "properties": {
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "proxy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Id"
          },
          "account_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Metadata"
          }
        },
        "type": "object",
        "title": "AccountUpdate"
      },
      "AccountWithLinksOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "cooldown_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldown Until"
          },
          "proxy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Id"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "tags": {
            "type": "string",
            "title": "Tags"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "total_usage_minutes": {
            "type": "number",
            "title": "Total Usage Minutes"
          },
          "usage_today_minutes": {
            "type": "number",
            "title": "Usage Today Minutes"
          },
          "usage_reset_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage Reset Date"
          },
          "device_links": {
            "items": {
              "$ref": "#/components/schemas/DeviceAccountOut"
            },
            "type": "array",
            "title": "Device Links",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "platform",
          "username",
          "display_name",
          "status",
          "cooldown_until",
          "proxy_id",
          "notes",
          "tags",
          "user_id",
          "created_at",
          "updated_at",
          "last_used_at",
          "total_usage_minutes",
          "usage_today_minutes",
          "usage_reset_date"
        ],
        "title": "AccountWithLinksOut",
        "description": "Detailed account response that includes the list of device links."
      },
...
      "AssignDeviceBody": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary",
            "default": false
          }
        },
        "type": "object",
        "required": ["device_id"],
        "title": "AssignDeviceBody",
        "description": "Used by POST /accounts/{account_id}/devices — assigns a device to an account."
      },
      "Body_bulk_import_csv_api_accounts_import_csv_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "CSV file with columns: platform, username, password, display_name, tags, notes"
          }
        },
        "type": "object",
        "required": ["file"],
        "title": "Body_bulk_import_csv_api_accounts_import_csv_post"
      },
      "BulkImportBody": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/BulkImportRow"
            },
            "type": "array",
            "title": "Accounts"
          }
        },
        "type": "object",
        "required": ["accounts"],
        "title": "BulkImportBody",
        "description": "JSON bulk import body."
      },
      "BulkImportResult": {
        "properties": {
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": ["created", "skipped", "total"],
        "title": "BulkImportResult"
      },
...
      "DeviceAccountOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          },
          "assigned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assigned At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "account_id",
          "is_primary",
          "assigned_at"
        ],
        "title": "DeviceAccountOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoundRobinBody": {
        "properties": {
          "account_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Account Ids"
          },
          "device_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Device Ids"
          }
        },
        "type": "object",
        "required": ["account_ids", "device_ids"],
        "title": "RoundRobinBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}